move cli dir to fil-tools for better go install name
