+++ exe/fil-tools/cmd/deal.go
