+++ exe/fil-tools/cmd/deals_store.go
