+++ exe/fil-tools/cmd/helpers.go
