+++ exe/fil-tools/cmd/init.go
