+++ exe/fil-tools/cmd/wallet_new.go
